Refactor project, both files and code
The biggest change to using PyTimer is that it no longer forces itself
to the top of the window stack.  Setting that attribute had an
unintended side-effect of putting the program in an unusable state if
it was minimized.  I was not able to find anything about why that was
the case, so it will have to lose that feature for the time being.

Arrange the files in a cleaner structure for later expansion, and give
them better names.  Also went in and rewrote a bit of the GUI code to
be more in line with the code I have written in more recent projects.

The py2app build is currently not working.