Add a chunk of work towards the new splits mode
The application can now open and display the simple split JSON-format
files that I wrote.  It can also save, but that functionality is not
yet attached to the interface.  The splitting is mostly there as well,
but that requires bringing together some things in the controller
first.

Some of the structure of the application was changed, like separating
the window and contents into pieces, adding some more shortcuts and
custom menubar.

I "fixed" a race condition that could be triggered when setting the
time while it was still running, by stopping the timer when editing.